[wip] Working on finding the right way to analyse things...
It is so odd to take the mean of samples that are not Gaussian. Why does this
lead to a Gamma distributed AFD? And why, when one takes the mean of the histogram itself
(i.e., the weighted mean), does this tend to make the lognormal AFD a better fit?
What is going on?